requirements.txt
anthropic==0.49.0
Flask==3.1.0
flask_cors==5.0.1
Pillow==11.1.0
protobuf==6.30.2
python-dotenv==1.1.0
Requests==2.32.3
google-genai
